meteo-weather-data/requirements.txt
--extra-index-url https://pkgs.dev.azure.com/quix-analytics/53f7fe95-59fe-4307-b479-2473b96de6d1/_packaging/public/pypi/simple/
quixstreams==2.9.0a
python-dotenv
pandas
requests